move config tests from model to kernel config
